requirements.txt
azure-core==1.19.0
azure-data-tables==12.1.0
azure-functions==1.7.2
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
isodate==0.6.0
msrest==0.6.21
numpy==1.21.2
oauthlib==3.1.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
